--- a/changelog/@unreleased/pr-49.v2.yml
+++ b/changelog/@unreleased/pr-49.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Reduce memory usage in ParamStorer utilities
+  links:
+  - https://github.com/palantir/witchcraft-go-params/pull/49