--- a/paramstorer.go
+++ b/paramstorer.go
@@ -1,7 +1,9 @@
 package wparams
 
 // ParamStorer is a type that stores safe and unsafe parameters. Keys should be unique across both SafeParams and
-// UnsafeParams (that is, if a key occurs in one map, it should not occur in the other).
+// UnsafeParams (that is, if a key occurs in one map, it should not occur in the other). For performance reasons,
+// the maps returned by SafeParams and UnsafeParams are references to the underlying storage and should not be modified
+// by the caller.
 type ParamStorer interface {
 	SafeParams() map[string]interface{}
 	UnsafeParams() map[string]interface{}
@@ -15,22 +17,11 @@ type ParamStorer interface {
 // example, if an unsafe key/value pair is provided by one ParamStorer and a later ParamStorer specifies a safe
 // key/value pair with the same key, the returned ParamStorer will store the last safe key/value pair).
 func NewParamStorer(paramStorers ...ParamStorer) ParamStorer {
-	safeParams := make(map[string]interface{})
-	unsafeParams := make(map[string]interface{})
+	collector := &mapParamStorer{}
 	for _, storer := range paramStorers {
-		if storer == nil {
-			continue
-		}
-		for k, v := range storer.SafeParams() {
-			safeParams[k] = v
-			delete(unsafeParams, k)
-		}
-		for k, v := range storer.UnsafeParams() {
-			unsafeParams[k] = v
-			delete(safeParams, k)
-		}
+		collector.copyFrom(storer)
 	}
-	return NewSafeAndUnsafeParamStorer(safeParams, unsafeParams)
+	return collector
 }
 
 // NewSafeParamStorer returns a new ParamStorer that stores the provided parameters as SafeParams.
@@ -38,25 +29,31 @@ func NewSafeParamStorer(safeParams map[string]interface{}) ParamStorer {
 	return NewSafeAndUnsafeParamStorer(safeParams, nil)
 }
 
+// NewSafeParam returns a new ParamStorer that stores a single unsafe parameter.
+func NewSafeParam(key string, value interface{}) ParamStorer {
+	return singleParamStorer{key: key, value: value, safe: true}
+}
+
 // NewUnsafeParamStorer returns a new ParamStorer that stores the provided parameters as UnsafeParams.
 func NewUnsafeParamStorer(unsafeParams map[string]interface{}) ParamStorer {
 	return NewSafeAndUnsafeParamStorer(nil, unsafeParams)
+}
+
+// NewUnsafeParam returns a new ParamStorer that stores a single unsafe parameter.
+func NewUnsafeParam(key string, value interface{}) ParamStorer {
+	return singleParamStorer{key: key, value: value, safe: false}
 }
 
 // NewSafeAndUnsafeParamStorer returns a new ParamStorer that stores the provided safe parameters as SafeParams and the
 // unsafe parameters as UnsafeParams. If the safeParams and unsafeParams have any keys in common, the key/value pairs in
 // the unsafeParams will be used (the conflicting key/value pairs provided by safeParams will be ignored).
 func NewSafeAndUnsafeParamStorer(safeParams, unsafeParams map[string]interface{}) ParamStorer {
-	storer := &mapParamStorer{
-		safeParams:   make(map[string]interface{}),
-		unsafeParams: make(map[string]interface{}),
-	}
+	storer := &mapParamStorer{}
 	for k, v := range safeParams {
-		storer.safeParams[k] = v
+		storer.putSafeParam(k, v)
 	}
 	for k, v := range unsafeParams {
-		storer.unsafeParams[k] = v
-		delete(storer.safeParams, k)
+		storer.putUnsafeParam(k, v)
 	}
 	return storer
 }
@@ -67,9 +64,82 @@ type mapParamStorer struct {
 }
 
 func (m *mapParamStorer) SafeParams() map[string]interface{} {
+	if m.safeParams == nil {
+		return map[string]interface{}{}
+	}
 	return m.safeParams
 }
 
 func (m *mapParamStorer) UnsafeParams() map[string]interface{} {
+	if m.unsafeParams == nil {
+		return map[string]interface{}{}
+	}
 	return m.unsafeParams
+}
+
+func (m *mapParamStorer) putSafeParam(k string, v interface{}) {
+	if m.safeParams == nil {
+		m.safeParams = map[string]interface{}{k: v}
+	} else {
+		m.safeParams[k] = v
+	}
+	delete(m.unsafeParams, k)
+}
+
+func (m *mapParamStorer) putUnsafeParam(k string, v interface{}) {
+	if m.unsafeParams == nil {
+		m.unsafeParams = map[string]interface{}{k: v}
+	} else {
+		m.unsafeParams[k] = v
+	}
+	delete(m.safeParams, k)
+}
+
+func (m *mapParamStorer) copyFrom(storer ParamStorer) {
+	if storer == nil {
+		return
+	}
+	// If this is one of our types, we can access the values directly and avoid intermediate map allocations.
+	switch st := storer.(type) {
+	case singleParamStorer:
+		if st.safe {
+			m.putSafeParam(st.key, st.value)
+		} else {
+			m.putUnsafeParam(st.key, st.value)
+		}
+	case *mapParamStorer:
+		for k, v := range st.safeParams {
+			m.putSafeParam(k, v)
+		}
+		for k, v := range st.unsafeParams {
+			m.putUnsafeParam(k, v)
+		}
+	default:
+		for k, v := range st.SafeParams() {
+			m.putSafeParam(k, v)
+		}
+		for k, v := range st.UnsafeParams() {
+			m.putUnsafeParam(k, v)
+		}
+	}
+}
+
+type singleParamStorer struct {
+	key   string
+	value interface{}
+	safe  bool
+}
+
+func (s singleParamStorer) SafeParams() map[string]interface{} {
+	if !s.safe {
+		return map[string]interface{}{}
+	}
+	return map[string]interface{}{s.key: s.value}
+}
+
+func (s singleParamStorer) UnsafeParams() map[string]interface{} {
+	if s.safe {
+		return map[string]interface{}{}
+	}
+	return map[string]interface{}{s.key: s.value}
 }