I saw some of this library popping up in memory profiles, so decided to take a look at reducing the number of map creations/allocations. 

This PR should have no functional changes, but makes the following optimizations:
* Defer allocating empty maps for safeParams/unsafeParams until we know we have something to populate
* Add `singleParamStorer` for cases we know only contain one parameter
* When merging storers, check the underlying type to see if we can avoid going through the public API

Old benchmark:
```
BenchmarkContextWithSafeParam
BenchmarkContextWithSafeParam-16              	  418755	      2710 ns/op	    2640 B/op	      30 allocs/op
BenchmarkContextWithSafeParams
BenchmarkContextWithSafeParams-16             	 1000000	      1036 ns/op	     880 B/op	      10 allocs/op
BenchmarkContextWithSafeAndUnsafeParam
BenchmarkContextWithSafeAndUnsafeParam-16     	  178443	      6517 ns/op	    6144 B/op	      63 allocs/op
BenchmarkContextWithSafeAndUnsafeParams
BenchmarkContextWithSafeAndUnsafeParams-16    	  510159	      2384 ns/op	    2048 B/op	      21 allocs/op
```

New benchmark:
```
BenchmarkContextWithSafeParam
BenchmarkContextWithSafeParam-16              	  855843	      1388 ns/op	    1488 B/op	      19 allocs/op
BenchmarkContextWithSafeParams
BenchmarkContextWithSafeParams-16             	 1465884	       831 ns/op	     832 B/op	       9 allocs/op
BenchmarkContextWithSafeAndUnsafeParam
BenchmarkContextWithSafeAndUnsafeParam-16     	  318692	      3596 ns/op	    3888 B/op	      42 allocs/op
BenchmarkContextWithSafeAndUnsafeParams
BenchmarkContextWithSafeAndUnsafeParams-16    	  603049	      1998 ns/op	    1904 B/op	      18 allocs/op
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/palantir/witchcraft-go-params/49)
<!-- Reviewable:end -->