@svc-autorelease 2021-01-14
Released [1.2.0](https://github.com/palantir/witchcraft-go-params/releases/tag/v1.2.0)